tests/wpt/web-platform-tests/websockets/Create-protocol-with-space.htm

<!DOCTYPE html>
<html>
<head>
    <title>W3C WebSocket API - Create WebSocket - protocol with space</title>
    <script type="text/javascript" src="/resources/testharness.js"></script>
    <script type="text/javascript" src="/resources/testharnessreport.js"></script>
    <script type="text/javascript" src="websocket.sub.js"></script>
</head>
<body>
    <div id="log"></div>
    <script type="text/javascript">
        test(function () {
            var wsocket;
            assert_throws("SYNTAX_ERR", function () { wsocket = CreateWebSocketWithSpaceInProtocol("ec ho") });
        }, "W3C WebSocket API - Create WebSocket - Pass a valid URL and a protocol string with a space in it - SYNTAX_ERR is thrown")
    </script>
</body>
</html>
